refactor(vite): isolate Vue feature options
