+++ npm/builder/vite/src/plugin/index.ts
@@ -39,6 +39,7 @@ import {
 } from "./vue-version.ts";
 import { resolveSharedConfig } from "./shared-config.ts";
 import * as configBridge from "./config-lifecycle.ts";
+import { resolveOptionsApiFlag } from "./vue-feature-defines.ts";
 
 export type { VizePluginState } from "./state.ts";
 
@@ -80,14 +81,6 @@ function resolveCompatibilityOptions(
   }
 
   return compatibility;
-}
-
-function parseDefine(value: unknown): unknown {
-  try {
-    return typeof value === "string" ? JSON.parse(value) : value;
-  } catch {
-    return value;
-  }
 }
 
 export function vize(options: VizeOptions = {}): Plugin[] {
@@ -134,13 +127,12 @@ export function vize(options: VizeOptions = {}): Plugin[] {
       patchCssModuleGenerateScopedName(userConfig);
 
       return {
-        // Vue 3 ESM bundler build requires these compile-time feature flags.
-        // @vitejs/plugin-vue normally provides them; vize must do so as its replacement.
+        // Vue 3 ESM bundler flags normally injected by @vitejs/plugin-vue.
         define: {
-          __VUE_OPTIONS_API__:
-            options.features?.optionsAPI ??
-            parseDefine(userConfig.define?.__VUE_OPTIONS_API__) ??
-            true,
+          __VUE_OPTIONS_API__: resolveOptionsApiFlag(
+            options.features?.optionsAPI,
+            userConfig.define?.__VUE_OPTIONS_API__,
+          ),
           __VUE_PROD_DEVTOOLS__: env.command === "serve",
           __VUE_PROD_HYDRATION_MISMATCH_DETAILS__: false,
         },
