+++ npm/builder/vite/src/vue-features.ts
@@ -1,0 +1,9 @@
+/** Vue runtime feature flags shared with `@vitejs/plugin-vue`. */
+export interface VizeVueFeatures {
+  /**
+   * Set to `false` to allow Vue's Options API code to be tree-shaken from
+   * production bundles.
+   * @default true
+   */
+  optionsAPI?: boolean;
+}
